Edit, juste a patch avec valgrind car fsanitize sa pue la merde

+++ Printf/srcs/basic/ft_getnbr_base.c
@@ -6,7 +6,7 @@
 /*   By: juspende <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/07/06 22:37:51 by juspende          #+#    #+#             */
-/*   Updated: 2017/11/25 16:33:48 by juspende         ###   ########.fr       */
+/*   Updated: 2017/11/26 14:15:16 by juspende         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -41,12 +41,12 @@ static int		my_error(char *base)
 	return (0);
 }
 
-char			*ft_getnbr_base(int nbr, char *base)
+char			*ft_getnbr_base(intmax_t nbr, char *base)
 {
 	char			str[4096];
-	unsigned int	i;
+	intmax_t		i;
 	int				x;
-	unsigned int	nb;
+	intmax_t		nb;
 
 	i = 0;
 	x = 0;
