+++ Printf/srcs/basic/ft_putint.c
@@ -6,16 +6,16 @@
 /*   By: juspende <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/07/04 15:06:32 by juspende          #+#    #+#             */
-/*   Updated: 2017/11/22 17:03:39 by juspende         ###   ########.fr       */
+/*   Updated: 2017/11/26 14:46:33 by juspende         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include <unistd.h>
 #include "../../include/ft_printf.h"
 
-void	ft_putint(int nb, t_flag *flag)
+void	ft_putint(intmax_t nb, t_flag *flag)
 {
-	unsigned int	i;
+	intmax_t	i;
 
 	if (nb < 0)
 	{
